feat(graphql): restructure hero-related GraphQL types and queries for improved organization
