+++ Heroes.Server/Hero/HeroRoleGraphType.cs
@@ -1,4 +1,4 @@
-﻿namespace Heroes.Server.Gql.Types;
+namespace Heroes.Server.Hero;
 
 public class HeroRoleGraphType : EnumerationGraphType
 {
