+++ Heroes.Server/HeroStat/HeroStatsGraphType.cs
@@ -1,4 +1,4 @@
-﻿namespace Heroes.Server.Gql.Types;
+namespace Heroes.Server.HeroStat;
 
 public class HeroStatsGraphType : ObjectGraphType<HeroStats>
 {
